add component index for exporting component parts
